# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customer 360 Intelligence Platform",
    "description": "\nCustomer 360 Intelligence Platform API.\n\nDeterministic customer data, agent-generated insights and grounded natural-language Q&A over a\nread-only banking dataset. Monetary values are integer cents internally and decimal strings at this\nboundary; every response carries a correlation ID and trace ID in `meta`.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/signals": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "The caller's ranked, entitlement-scoped signals worklist",
        "description": "The prioritized daily worklist across the caller's entitled book (task 17.3, 17.5).\n\nRanked by severity x value-at-stake x recency, suppressing dismissed and cooled-off signals.\nReturns an empty page when signal detection has not run yet.",
        "operationId": "list_signals_signals_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Page_dict_str__Any___"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customer_id}/signals": {
      "get": {
        "tags": [
          "signals"
        ],
        "summary": "Signals for one customer (drill-down)",
        "description": "Every signal for one customer, behind the same 403-vs-404 gate as the 360 view (task\n17.5).",
        "operationId": "list_customer_signals_customers__customer_id__signals_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_dict_str__Any__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/signals/{signal_id}/dismiss": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Dismiss a signal for the calling user",
        "description": "Suppress a signal from the caller's queue for the cooling-off window (task 17.5).",
        "operationId": "dismiss_signal_signals__signal_id__dismiss_post",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DismissResult_"
                }
              }
            }
          },
          "404": {
            "description": "No such signal, or not entitled to its customer"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/signals/{signal_id}/ack": {
      "post": {
        "tags": [
          "signals"
        ],
        "summary": "Acknowledge (mark actioned) a signal for the calling user",
        "description": "Mark a signal actioned for the caller, permanently suppressing it from their queue (task\n17.5).",
        "operationId": "ack_signal_signals__signal_id__ack_post",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DismissResult_"
                }
              }
            }
          },
          "404": {
            "description": "No such signal, or not entitled to its customer"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DismissResult": {
        "properties": {
          "signal_id": {
            "type": "integer",
            "title": "Signal Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "signal_id",
          "status"
        ],
        "title": "DismissResult",
        "description": "The outcome of a dismiss/ack write."
      },
...
      "Envelope_DismissResult_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DismissResult"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[DismissResult]"
      },
...
      "Envelope_Page_dict_str__Any___": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Page_dict_str__Any__"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[Page[dict[str, Any]]]"
      },
...
      "Envelope_dict_str__Any__": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[dict[str, Any]]"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNAUTHENTICATED",
          "ENTITLEMENT_DENIED",
          "CUSTOMER_NOT_FOUND",
          "VALIDATION_ERROR",
          "AGENT_TIMEOUT",
          "UPSTREAM_UNAVAILABLE",
          "RATE_LIMITED",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "description": "The complete error code set from design \u00a76.3."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Meta": {
        "properties": {
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "masked_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Masked Fields"
          },
          "computed_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Computed Fields"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ModuleError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "correlation_id",
          "as_of"
        ],
        "title": "Meta",
        "description": "Envelope metadata."
      },
      "ModuleError": {
        "properties": {
          "module": {
            "type": "string",
            "title": "Module"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "module",
          "code",
          "message"
        ],
        "title": "ModuleError",
        "description": "A single domain module's failure, surfaced without failing the whole response."
      },
...
      "Page_dict_str__Any__": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "title": "Page[dict[str, Any]]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}